pyproject.toml
[project]
name = "modulo-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.14",
    "aiosqlite>=0.19.0",
    "chromadb==0.4.22",
    "fastapi>=0.128.0",
    "numpy<2.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pypdf>=3.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "tornado<6.2",
    "uvicorn>=0.40.0",
]
